refactor - ScriptResult and ScriptExecution : move to Rest package
