+++ core/java/iesi-rest-without-microservices/src/main/java/io/metadew/iesi/server/rest/scriptExecutionDto/inputParameter/InputParametersDto.java
@@ -1,4 +1,4 @@
-package io.metadew.iesi.server.rest.script.execution.inputParameter;
+package io.metadew.iesi.server.rest.scriptExecutionDto.inputParameter;
 
 import lombok.AllArgsConstructor;
